requirements.txt
# Core dependencies
pandas>=2.1.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.7.0
pytz>=2023.3

# Technical indicators
ta>=0.11.0

# Logging and utilities
colorlog>=6.8.0

# Broker API - KiteConnect (use latest for Python 3.12 compatibility)
kiteconnect>=4.2.0

# For Kite token server (only needed for get_kite_token.py)
flask>=3.0.0

# For backtesting
matplotlib>=3.8.0
tabulate>=0.9.0